Fixed oversight that causes heap corruption
Fixed an oversight where initializing an array with dynamic_array_init(void*) and then pushing into it (causing an allocation on push) would not properly increase the capacity of the dynamic array (cap *= 2, which means that capacity would always stay at 0 when using init or alloc with a size of 0), which would later lead to heap corruption when pushing elements over the capacity of the vector or when calling free.

Now, the dynamic arrays perform an initial allocation of 8 bytes if the capacity is 0 when performing the first push operation on an initialized dynamic array.